First commit in a while. Basic functionality for a GUI database editor has been added, but it is still very early in development and many parts may be inefficiently implemented.
The CLI is not being actively developed as the GUI is much more efficient at the moment until certain algorithms are implemented to make it easier.

This includes all of what I think is required to run an IntelliJ IDEA version of the code on my system, but there may be missing elements. IDEA was chosen as the IDE because it seems to be standard. If this is an unnecessary limitation or if there are things I can do to make it easier to open in IDEA, please let me know.

I've organized various parts of the project into different packages under the project package net.openpatterns.ents. The essential package is database_library which contains the use-independent code for using and Ents Database, and the other packages are for manipulating databases. Implementing interfaces within the database_library can bypass these 1st party user interfaces.

The main user interface is in the gui package. The singleton Launcher handles loading of databases into Explorer instances. Right now it can only load an example hard-coded database which can be used to test the Explorer GUI.

Explorer GUI has minimal functionality like deleting an ent/node. Adding functionality to this is the current task todo in order to manually manipulate hierarchies. Smart systems like the Sorter/Stream need populated databases to be useful, so the basic Explorer is more important.

If package or code structure is improper please let me know! I've separated packages for organizational reasons primarily, and hopefully it doesn't impede usage.

I've tried to document the code sufficiently, including package-info.java files within the packages, but it's exhaustive yet.